upd: use double instead of float

+++ dependencies/android/src/main/java/org/haxe/extension/ExtensionAdjust.java
@@ -59,7 +59,7 @@ public class ExtensionAdjust extends Extension {
 		Log.i(TAG, eventName + " tracked");
 	}
 
-	public static void trackRevenueEvent(String eventName, float value, String currency) {
+	public static void trackRevenueEvent(String eventName, double value, String currency) {
 		AdjustEvent adjustEvent = new AdjustEvent(eventName);
 		adjustEvent.setRevenue(value, currency);
 		Adjust.trackEvent(adjustEvent);
